test(act-rules): migrate to TS and clean up (+ document) code
